# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Fortune {
  id: number;
  text: string;
  category: string;
  luckyNumbers: number[];
  date: string;
}

export interface FortuneCategory {
  name: string;
  count: number;
  emoji: string;
}

export interface WeeklyFortune {
  id: number;
  text: string;
  category: string;
  luckyNumbers: number[];
  weekNumber: number;
  weekLabel: string;
  expiresAt: string;
}

export type VerifyPaymentRequestMathChallenge = {
  a?: number;
  b?: number;
  text?: string;
};

export interface VerifyPaymentRequest {
  amount: string;
  method: string;
  note: string;
  mathAnswer: number;
  mathChallenge: VerifyPaymentRequestMathChallenge;
}

export interface VerifyPaymentAck {
  ok: boolean;
  message: string;
  pending?: boolean;
  token?: string;
}

export type TokenStatusStatus = typeof TokenStatusStatus[keyof typeof TokenStatusStatus];


export const TokenStatusStatus = {
  pending: 'pending',
  approved: 'approved',
  expired: 'expired',
  invalid: 'invalid',
} as const;

export interface TokenStatus {
  status: TokenStatusStatus;
}

export type ChimePaymentMathChallenge = {
  a: number;
  b: number;
  text: string;
};

export interface ChimePayment {
  qrCodeDataUrl: string;
  chimeTag: string;
  amount: string;
  instructions: string;
  mathChallenge: ChimePaymentMathChallenge;
}

export interface MathChallenge {
  a: number;
  b: number;
  text: string;
}

export type GetRandomFortuneParams = {
category?: GetRandomFortuneCategory;
};

export type GetRandomFortuneCategory = typeof GetRandomFortuneCategory[keyof typeof GetRandomFortuneCategory];


export const GetRandomFortuneCategory = {
  Love: 'Love',
  Career: 'Career',
  Health: 'Health',
  Wealth: 'Wealth',
  Wisdom: 'Wisdom',
} as const;

export type CheckPaymentTokenParams = {
token: string;
};

export type GetChimeQrCodeParams = {
amount: GetChimeQrCodeAmount;
};

export type GetChimeQrCodeAmount = typeof GetChimeQrCodeAmount[keyof typeof GetChimeQrCodeAmount];


export const GetChimeQrCodeAmount = {
  '099': '0.99',
  '499': '4.99',
} as const;