# lib/api-zod/src/generated/types/chimePayment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */
import type { ChimePaymentMathChallenge } from './chimePaymentMathChallenge';

export interface ChimePayment {
  qrCodeDataUrl: string;
  chimeTag: string;
  amount: string;
  instructions: string;
  mathChallenge: ChimePaymentMathChallenge;
}