# lib/api-zod/src/generated/types/fortune.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */

export interface Fortune {
  id: number;
  text: string;
  category: string;
  luckyNumbers: number[];
  date: string;
}