# lib/api-zod/src/generated/types/mathChallenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */

export interface MathChallenge {
  a: number;
  b: number;
  text: string;
}