# lib/api-zod/src/generated/types/tokenStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */
import type { TokenStatusStatus } from './tokenStatusStatus';

export interface TokenStatus {
  status: TokenStatusStatus;
}