# lib/api-zod/src/generated/types/verifyPaymentAck.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyPaymentAck {
  ok: boolean;
  message: string;
  pending?: boolean;
  token?: string;
}