# lib/api-zod/src/generated/types/verifyPaymentRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */
import type { VerifyPaymentRequestMathChallenge } from './verifyPaymentRequestMathChallenge';

export interface VerifyPaymentRequest {
  amount: string;
  method: string;
  note: string;
  mathAnswer: number;
  mathChallenge: VerifyPaymentRequestMathChallenge;
}