# lib/api-zod/src/generated/types/weeklyFortune.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mystic Cookie API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyFortune {
  id: number;
  text: string;
  category: string;
  luckyNumbers: number[];
  weekNumber: number;
  weekLabel: string;
  expiresAt: string;
}